[#6] chore: 용도에 따라 webpack 설정 파일 분리 및 파비콘 파일 변경
