+++ frontend/config/resolvePathFromRoot.js
@@ -1,0 +1,6 @@
+const path = require('path');
+
+const resolvePathFromRoot = (...pathSegments) =>
+  path.resolve(__dirname, '..', ...pathSegments);
+
+module.exports = resolvePathFromRoot;
