+++ frontend/config/webpack.config.js
@@ -1,16 +1,10 @@
-const { HotModuleReplacementPlugin } = require('webpack');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
-const { CleanWebpackPlugin } = require('clean-webpack-plugin');
 
-const path = require('path');
+const resolvePathFromRoot = require('./resolvePathFromRoot');
 
-const MODE = { prod: 'production', dev: 'development' };
 const PUBLIC_PATH = '/';
-const resolvePathFromRoot = (...pathSegments) =>
-  path.resolve(__dirname, '..', ...pathSegments);
 
-const config = {
-  mode: MODE.dev,
+module.exports = {
   target: 'web',
   resolve: {
     extensions: ['.js', '.jsx', '.ts', '.tsx'],
@@ -30,21 +24,6 @@ const config = {
     path: resolvePathFromRoot('dist'),
     filename: '[name].[hash].js',
     publicPath: PUBLIC_PATH,
-  },
-  devtool: 'source-map',
-  devServer: {
-    port: 4000,
-    inline: true,
-    compress: true,
-    hot: true,
-    proxy: {
-      '/api': {
-        target: 'http://localhost:3000',
-      },
-    },
-    historyApiFallback: {
-      rewrites: [{ from: /^\/*$/, to: '/index.html' }],
-    },
   },
   module: {
     rules: [
@@ -74,20 +53,10 @@ const config = {
     ],
   },
   plugins: [
-    new CleanWebpackPlugin(),
-    new HotModuleReplacementPlugin(),
     new HtmlWebpackPlugin({
-      favicon: './src/assets/favicon.png',
       filename: 'index.html',
       template: resolvePathFromRoot('src', 'index.html'),
+      favicon: resolvePathFromRoot('src', 'assets', 'favicon.ico'),
     }),
   ],
 };
-
-if (process.env.NODE_ENV === MODE.prod) {
-  config.mode = MODE.prod;
-  delete config.devtool;
-  delete config.devServer;
-}
-
-module.exports = config;
